build packages in temporary directories
If we build them all from the project root, the results will all go in the
project root (which is a bit of a problem if the built binary has the same name
as the package's directory).

We could just CD into the package's directory but it's safer to just build all
this stuff in a temporary directory.